+++ imageZIP/__init__.py
@@ -1,1 +1,10 @@
+# -*- coding: utf-8 -*-
+"""
+Archive (encrypt) files and directories to image file
+"""
 from .imageZIP import zip, unzip
+__version__ = '3'
+__author__ = 'Herman Schechkin (hermanTenuki)'
+__copyright__ = 'Copyright (c) 2020 Herman Schechkin'
+__license__ = 'MIT'
+__email__ = 'itseasy322@gmail.com'