+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 Change Log
 =========================
 
+3 (27.11.2020)
+-------------------------
+- Linux path fix for an output image file;
+- New color modes: "rainbow", "red", "blue";
+- "heat_map" and "heat_map_toxic" color modes is now can be used with more shorter "heat" and "heat_toxic" names;
+- Docstrings.
+- Other minor fixes.
+
 2.1 (25.11.2020)
 -------------------------
 - Small linux fix.